x86/paravirt: don't enter/leave lazy mode in interrupts.
We already ignore the current state of lazy mode in interrupts, but we
should also ignore any attempt to enter/leave lazy mode within
an interrupt context.

enter_lazy() will BUG if it sees an attempt at a nested entry to lazy
mode, which is generally an error.  However, it's possible that an
interrupt handler may do something that would trigger a batched MMU
update, for example, and that could interrupt an existing batched update.

Signed-off-by: Jeremy Fitzhardinge <jeremy.fitzhardinge@citrix.com>
Reported-by: Chuck Anderson <chuck.anderson@oracle.com>
Cc: Jan Beulich <JBeulich@novell.com>
Cc: Stable Kernel <stable@kernel.org>

Signed-off-by: Cristian Gafton <gafton@amazon.com>
Conflicts:
	arch/x86/kernel/paravirt.c
X-Patch-Url: http://lists.xen.org/archives/html/xen-devel/2010-12/msg00505.html