demodularize  - only support the linux submodule for now
it is getting harder to know what is the right thing to do for stuff like
patch application and generation. Do thos patches get generated from each
module for the same spec file? or for multiple spec files? Until we have a
clear answer on how this will get used, we keep things simpler and
straightforward