virtio-pci: also bind to Amazon PCI vendor ID
See https://issues.oasis-open.org/browse/VIRTIO-16 although it
was prematurely closed.

It is not possible to produce Windows drivers for virtio that can
be published in Windows update because Red Hat controls the vendor ID
that virtio is using and only provides non-redistributable drivers for
Windows.

The only option right now is to expose a virtio device with a
different vendor ID if you want to support Windows with virtio.

Cc: Matt Wilson <msw@amazon.com>
Cc: Rusty Russell <rusty@rustcorp.com.au>
Cc: Michael Tsirkin <mst@redhat.com>
Signed-off-by: Anthony Liguori <aliguori@amazon.com>

Conflicts:
	drivers/virtio/virtio_pci.c