drivers/amazon: xen-blkfront: resurrect per-device lock
In requset-based mode, a spin lock in the first blkfront_ring_info is
also used as a queue lock. However, the structure is freed and re-created
while resuming from Xen suspend. Therefore, the block layer may still
continue to use the freed memory as a queue lock. Obviously, it can
result in crash, hang or corruption.

Not to make the queue lock stale, resurrect per-device (vbd) lock in
blkfront_info which is never freed during Xen suspend and use it in
request-based mode. This is bascially the same as what the driver was
doing until commit 11659569f720 ("xen/blkfront: split per device
io_lock"). If the driver is in blk-mq mode, just use the lock(s) in
blkfront_ring_info.

Reported-by: Imre Palik <imrep@amazon.com>
Reported-and-tested-by: Thomas Friebel <friebelt@amazon.com>
Signed-off-by: Munehisa Kamata <kamatam@amazon.com>
Reviewed-by: Eduardo Valentin <eduval@amazon.com>
Reviewed-by: Guru Anbalagane <guruanb@amazon.com>

CR: https://cr.amazon.com/r/7475903/