config: make TCP Cubic congestion algorithm default
We enabled TCP BBR congestion algorithm in commit 80bc03b156c4 ("config:
enable Bottleneck Bandwidth and RTT by default") and made it the default
algorithm instead of TCP Cubic. However, it causes noticeable
performance regression even in simple bandwidth benchmark using netperf.

This is probably a known behavior in TCP BBR and the benchmark is a kind
of micro benchmark, but we're not very clear how much this can affect
real-world applications or workloads.

To avoid causing unexpected regression in the field, make TCP Cubic
default back. For those who prefer TCP BBR, we keep it as built-in, so
it can be configured via tradtional sysctl.

Fixes: 80bc03b156c4 ("config: enable Bottleneck Bandwidth and RTT by default")
Signed-off-by: Munehisa Kamata <kamatam@amazon.com>
Reviewed-by: iliana weller <iweller@amazon.com>
Reviewed-by: Guru Anbalagane <guruanb@amazon.com>
Reviewed-by: Alakesh Haloi <alakeshh@amazon.com>

CR: https://cr.amazon.com/r/7761213/