drivers/amazon: xen-blkfront: ensure no reqs/rsps in rings before disconnecting
When disconnecting the frontend from the backend in blkfront_freeze(),
there still may be unconsumed requests or responses in the rings,
especially when the backend is backed by network-based device. If the
frontend gets disconnected with such reqs/rsps remaining there, it can
cause grant warnings and/or losing reqs/rsps by freeing pages afterward.
This can lead resumed kernel into unrecoverable state like unexpected
freeing of grant page and/or hung task due to the lost reqs or rsps.
Therefore we have to ensure that there is no unconsumed requests or
responses before disconnecting.

Actually, the frontend just needs to wait for some amount of time so that
the backend can process the requests, put responses and notify the
frontend back. Timeout used here is based on some heuristic. If we somehow
hit the timeout, it would mean something serious happens in the backend,
the frontend will just return an error to PM core and PM
suspend/hibernation will be aborted.

This may be something should be fixed by the backend side, but a frontend
side fix is probably still worth doing to work with broader backends.

Fixes: c57e22999619 ("drivers/amazon: xen-blkfront: add callbacks for PM suspend and hibernation")
Signed-off-by: Munehisa Kamata <kamatam@amazon.com>
Reviewed-by: Vallish Vaidyeshwara <vallish@amazon.com>
Reviewed-by: Eduardo Valentin <eduval@amazon.com>

CR: https://cr.amazon.com/r/7824995/