configs: split hardware specific configs
As per review done with principal (gafton@), karoo
kernel must be VM tuned. This means several options
that are in karoo kernel does not make sense for VM,
and are specific for Baremetal This patch attempts
to mitigate this by moving configs that make sense
to baremetal to specific files and rolling back to
VM tuned configs. The configs for baremetal are saved
for the case we need to use this kernel in baremetal.

Note that this patch also refreshes the current configs
with the output of make reconfig-prep and make reconfig,
as the tree was left with several configs not properly
selected.

Fixes: 80aeb23 config: Enable baremetal drivers
Fixes: 4308ab1 config: enable NVDIMM and PMEM drivers
Fixes: 80aeb23 config: Enable baremetal drivers
Reviewed-by: Alakesh Haloi <alakeshh@amazon.com>
Reviewed-by: Anchal Agarwal <anchalag@amazon.com>
Suggested-by: Cristian Gafton <gafton@amazon.com>
Signed-off-by: Eduardo Valentin <eduval@amazon.com>

CR: https://cr.amazon.com/r/7883729/
Signed-off-by: Eduardo Valentin <eduval@amazon.com>