config changes to enable FIPS boot
Add kernel configs to enable cryptographic modules used by crypto tests from
tcrypt driver. When FIPS mode is enabled in a system having the package
dracut-fips, it tries to check integrity ofcryptographic algorithms enabled
in testmgr. If FIPS is enabled for an algorithm, then the kernel module used
by the algorithm needs to be loaded. Scripts installed with dracut-fips package
handles this. With these configs enabled, we make sure that the kernel boots in
fips mode and all integrity verification for the enabled tests succeeds.

When kernel is not running in FIPS mode (i.e. kernel cmdline do not have
fips=1), then dracut-fips scripts are not run and there is no self-test run
in that case.

Signed-off-by: Alakesh Haloi <alakeshh@amazon.com>
Reviewed-by: Eduardo Valentin <eduval@amazon.com>
Reviewed-by: Anchal Agarwal <anchalag@amazon.com>
CR: https://cr.amazon.com/r/7891817/