add acpi spcr table parser
Config to allow OS use ACPI SPCR table to get console information by parsing it.
This feature is enabled to help bare metal instance debugging via early console.
It is necessary to fix console issues that we had on nitro instances (i3.metal).
The issue is described in detail in the tt: https://tt.amazon.com/0130202057.

The 8250 driver allocates an array of CONFIG_SERIAL_8250_RUNTIME_UARTS ports
and allocates COM1 to COM4 at the first 4 of the array. It thenprobes the
registers at ISA I/O ports and figures out which ones are enabled. (Only COM2
is enabled for i3.metal.) When the 8250 PCI device gets registered, it first
searches for an empty entry and falls back to a disabled device's entry when
the array is all occupied. This results in different naming for the PCI device
depending on the config option.

When CONFIG_SERIAL_8250_RUNTIME_UARTS > 4, it picks up the fifth entry. This
becomes ttyS4. When CONFIG_SERIAL_8250_RUNTIME_UARTS <= 4, it picks up the
first entry (COM1), so this becomes ttyS0. In summary, we want to avoid this
issue, by letting the OS use ACPI SPCR table to get console information for
early console, by parsing it. ACPI SPCR parsing patch is not yet available in
4.9 stable trees and has been cherry picked from mainline.

Signed-off-by: Alakesh Haloi <alakeshh@amazon.com>
Reviewed-by: Vallish Vaidyeshwara <vallish@amazon.com>
Reviewed-by: Guru Anbalagane <guruanb@amazon.com>
CR: https://cr.amazon.com/r/8313452/