netfilter: xt_conncount: Fix underflow in connection counting
In commit 414c083eeb29, we fixed an issue of connection counting by
introducing a pending_add flag to each connection tracking struct.
However, clearing the pending_add does not seem to happen on time
causing some of the connections in the data structure maintained by
connlimit module, not to be cleared. This causes available connections
to be reduced. The fix is not use pending_add flag and instead use a
time stamp and cpu core id for each connection created, as done in
this upstream patch http://patchwork.ozlabs.org/patch/932399/. This
seems to fix  the connection counting race without causing any leak.
If we see a connection in the xt_connlimit but not in netfilter's
conntrack table, and see that the connection was created less than
2 jiffies again on a different core, then we do not delete this
connection, as it is still in the process of making it to the nf
connection tracking table.

Signed-off-by: Alakesh Haloi <alakeshh@amazon.com>
Fixes: 414c083eeb29f73dc68c2ba0813a20d2f9856555
Reviewed-by: Frank van der Linden <fllinden@amazon.com>
Reviewed-by: Vallish Vaidyeshwara <vallish@amazon.com>
Tested-by: Hazem Mohamed Abuelfotoh <abuehaze@amazon.com>
CR: https://code.amazon.com/reviews/CR-4301474