[AL1] Rebase to v4.14.94
Kernel rebase to v4.14.94. As a part of the rebase, we need to
enable CONFIG_CIFS_ALLOW_INSECURE_LEGACY. This config was added to
4.14 as a dependency for CONFIG_CIFS_POSIX but 4.14.94 missed the
commit that actually added this new config. Without this config
being enabled, CONFIG_CIFS_POSIX will get disabled by make reconfig
which will break compatibility. We had to cherry pick the following
commit to keep things in check.

04f67ed3d83aa  cifs: allow disabling insecure dialects in the config

This does not change anything in practice for our kernel, as this new
config is a result of refactoring existing code. In spite of the name,
it does not add any new vulnerability. It justs marks few older versions
of samba protocol as insecure.

Signed-off-by: Alakesh Haloi <alakeshh@amazon.com>
Reviewed-by: Anchal Agarwal <anchalag@amazon.com>
Reviewed-by: Frank van der Linden <fllinden@amazon.com>
Reviewed-by: Qian Lu <luqia@amazon.com>
Reviewed-by: Vallish Vaidyeshwara <vallish@amazon.com>
CR: https://code.amazon.com/reviews/CR-5194640