lustre: silence printk format warnings about timespec.tv_sec
tv_sec may be defined as long or long long depending on timespec64
availability and depending on whether this is a 64bit kernel or not.

This leads to printk format warnings. Silence them by always casting
the value to unsigned long - the value we're printing here is not
larger than 32 bits in any case.

Signed-off-by: Frank van der Linden <fllinden@amazon.com>